requirements.txt
streamlit
pandas
plotly
openai-agents
PyPDF2
python-dotenv